+++ cmsafops/R/mon_num_above.R
@@ -1,0 +1,62 @@
+#'Number of timesteps per month above a threshold.
+#'
+#'This function counts the number of timesteps above a certain threshold for each 
+#'month and grid point of a dataset. This operator should be applied to data with
+#'temporal resolution < monthly (e.g., daily).
+#'
+#'@param var Name of NetCDF variable (character).
+#'@param infile Filename of input NetCDF file. This may include the directory
+#'  (character).
+#'@param outfile Filename of output NetCDF file. This may include the directory
+#'  (character).
+#'@param nc34 NetCDF version of output file. If \code{nc34 = 3} the output file will be
+#'  in NetCDFv3 format (numeric). Default output is NetCDFv4.
+#'@param overwrite logical; should existing output file be overwritten?
+#'@param verbose logical; if TRUE, progress messages are shown
+#'
+#'@return A NetCDF file including a time series of monthly maxima is written.
+#'@export
+#'
+#'@family monthly statistics
+#'
+#' @examples
+#'## Create an example NetCDF file with a similar structure as used by CM
+#'## SAF. The file is created with the ncdf4 package.  Alternatively
+#'## example data can be freely downloaded here: <https://wui.cmsaf.eu/>
+#'
+#'library(ncdf4)
+#'
+#'## create some (non-realistic) example data
+#'
+#'lon <- seq(5, 15, 0.5)
+#'lat <- seq(45, 55, 0.5)
+#'time <- seq(as.Date("2000-01-01"), as.Date("2000-03-31"), "days")
+#'origin <- as.Date("1983-01-01 00:00:00")
+#'time <- as.numeric(difftime(time, origin, units = "hour"))
+#'data <- array(250:350, dim = c(21, 21, 91))
+#'
+#'## create example NetCDF
+#'
+#'x <- ncdim_def(name = "lon", units = "degrees_east", vals = lon)
+#'y <- ncdim_def(name = "lat", units = "degrees_north", vals = lat)
+#'t <- ncdim_def(name = "time", units = "hours since 1983-01-01 00:00:00",
+#'  vals = time, unlim = TRUE)
+#'var1 <- ncvar_def("SIS", "W m-2", list(x, y, t), -1, prec = "short")
+#'vars <- list(var1)
+#'ncnew <- nc_create(file.path(tempdir(),"CMSAF_example_file.nc"), vars)
+#'ncvar_put(ncnew, var1, data)
+#'ncatt_put(ncnew, "lon", "standard_name", "longitude", prec = "text")
+#'ncatt_put(ncnew, "lat", "standard_name", "latitude", prec = "text")
+#'nc_close(ncnew)
+#'
+#'## Determine the monthly number of timesteps above a threshold of the example 
+#'## CM SAF NetCDF file and write the output to a new file.
+#'mon_num_above(var = "SIS", thld = 300, infile = file.path(tempdir(),"CMSAF_example_file.nc"), 
+#'  outfile = file.path(tempdir(),"CMSAF_example_file_mon_num_above.nc"))
+#'
+#'unlink(c(file.path(tempdir(),"CMSAF_example_file.nc"), 
+#'  file.path(tempdir(),"CMSAF_example_file_mon_num_above.nc")))
+mon_num_above <- function(var, thld = 0, infile, outfile, nc34 = 4, overwrite = FALSE, 
+                          verbose = FALSE) {
+  mon_num_wrapper(1, var, thld, infile, outfile, nc34, overwrite, verbose)
+}