+++ cmsaf/NEWS.md
@@ -2,7 +2,8 @@
 
 ## 3.1.2
 
-- Add new operators timavg, num_above, num_below, num_equal
+- Add new operators timavg, num_above, num_below, num_equal,
+  mon_num_above, mon_num_below, mon_num_equal
 
 ## 3.1.1